use nanosleep instead of usleep for linux

+++ kernel/src/arch/aarch64/delay.rs
@@ -9,11 +9,11 @@ impl Delay for ArchDelay {
         unsafe { core::arch::asm!("wfi") };
     }
 
-    fn wait_microsec(usec: u32) {
+    fn wait_microsec(usec: u64) {
         let frq = cpu::cntfrq_el0::get();
         let t = cpu::cntpct_el0::get();
 
-        let end = t + ((frq / 1000) * usec as u64) / 1000;
+        let end = t + ((frq / 1000) * usec) / 1000;
 
         while cpu::cntpct_el0::get() < end {
             cpu::isb();
