+++ kernel/src/arch/x86_64/delay.rs
@@ -7,7 +7,7 @@ impl Delay for ArchDelay {
         unsafe { core::arch::asm!("hlt") };
     }
 
-    fn wait_microsec(usec: u32) {
+    fn wait_microsec(usec: u64) {
         super::acpi::wait_usec(usec);
     }
 }
